Update suffix test post-suffix module move
WIP #2127
